refactor: improve package naming for clarity
- Rename internal/server → internal/artifactserver
  (serves temp artifacts for Playwright, not the SaaS server)
- Rename internal/web → internal/dashboard
  (local CLI dashboard, distinct from web/ SvelteKit frontend)
- Move mock_provider.go → mock_provider_test.go
  (mocks should not compile into production binary)

This clarifies the distinction between:
- internal/api: SaaS API server
- internal/artifactserver: local temp file server for CLI
- internal/dashboard: local web UI for CLI (`heisenberg serve`)
- web/: SvelteKit SaaS frontend
