+++ internal/artifactserver/server.go
@@ -1,4 +1,4 @@
-package server
+package artifactserver
 
 import (
 	"context"
