+++ internal/artifactserver/server_test.go
@@ -1,4 +1,4 @@
-package server
+package artifactserver
 
 import (
 	"io"
