+++ internal/dashboard/handler.go
@@ -1,4 +1,4 @@
-package web
+package dashboard
 
 import (
 	"embed"
