+++ internal/dashboard/sse.go
@@ -1,4 +1,4 @@
-package web
+package dashboard
 
 import (
 	"encoding/json"
