+++ internal/dashboard/static/index.html
